requirements.txt
Django
Pillow
MySQL-python
django-crispy-forms==1.6.0
python-dateutil
